[UserBundle][CoreBundle] Move UserProvider to Core
