--- src/Sylius/Bundle/CustomerBundle/Form/EventSubscriber/CustomerRegistrationFormSubscriber.php
@@ -1,74 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Sylius\Bundle\CustomerBundle\Form\EventSubscriber;
-
-use Sylius\Component\Resource\Repository\RepositoryInterface;
-use Sylius\Component\Customer\Model\CustomerInterface;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\Form\Exception\UnexpectedTypeException;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
-
-/**
- * @author Michał Marcinkowski <michal.marcinkowski@lakion.com>
- */
-class CustomerRegistrationFormSubscriber implements EventSubscriberInterface
-{
-    /**
-     * @var RepositoryInterface
-     */
-    protected $customerRepository;
-
-    /**
-     * @param RepositoryInterface $customerRepository
-     */
-    public function __construct(RepositoryInterface $customerRepository)
-    {
-        $this->customerRepository = $customerRepository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedEvents()
-    {
-        return [
-            FormEvents::PRE_SUBMIT => 'preSubmit',
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function preSubmit(FormEvent $event)
-    {
-        $rawData = $event->getData();
-        $form = $event->getForm();
-        $data = $form->getData();
-
-        if (!$data instanceof CustomerInterface) {
-            throw new UnexpectedTypeException($data, CustomerInterface::class);
-        }
-
-        // if email is not filled in, go on
-        if (!isset($rawData['email']) || empty($rawData['email'])) {
-            return;
-        }
-        $existingCustomer = $this->customerRepository->findOneBy(['email' => $rawData['email']]);
-        if (null === $existingCustomer || null !== $existingCustomer->getUser()) {
-            return;
-        }
-
-        $existingCustomer->setUser($data->getUser());
-        $form->setData($existingCustomer);
-    }
-}
